Version 0.015 (12 Mayo 2017)
Incluye la consulta de hospitales por ciudad y latitud-longitud.

Incluye la modificación de la consulta hospitales para hacerla en un
solo método del ServiceController (más eficiente).

+++ CMH-Beta/src/mx/com/stsystems/cmh/beta/dto/Hospital.java
@@ -18,9 +18,9 @@ public class Hospital implements Serializable {
 	@JsonIgnore
 	private int idEstado;
 	@JsonIgnore
-	private float latitud;
+	private double latitud;
 	@JsonIgnore
-	private float longitud;
+	private double longitud;
 
 	public String getIdHospital() {
 		return idHospital;
@@ -70,19 +70,19 @@ public class Hospital implements Serializable {
 		this.idEstado = idEstado;
 	}
 
-	public float getLatitud() {
+	public double getLatitud() {
 		return latitud;
 	}
 
-	public void setLatitud(float latitud) {
+	public void setLatitud(double latitud) {
 		this.latitud = latitud;
 	}
 
-	public float getLongitud() {
+	public double getLongitud() {
 		return longitud;
 	}
 
-	public void setLongitud(float longitud) {
+	public void setLongitud(double longitud) {
 		this.longitud = longitud;
 	}
